address new review comments.
